test(auditlog): keep export regression in handler tests
Amp-Thread-ID: https://ampcode.com/threads/T-01a074e2-49a2-7038-91cd-a94d6932d2b3
